+++ server/impl/src/main/java/com/walmartlabs/concord/server/process/queue/NoneWaitHandler.java
@@ -20,13 +20,20 @@ package com.walmartlabs.concord.server.process.queue;
  * =====
  */
 
+import com.walmartlabs.concord.server.sdk.ProcessStatus;
+
 import javax.inject.Named;
 import javax.inject.Singleton;
+import java.util.Arrays;
+import java.util.HashSet;
+import java.util.Set;
 import java.util.UUID;
 
 @Named
 @Singleton
 public class NoneWaitHandler implements ProcessWaitHandler<NoneCondition> {
+
+    private static final Set<ProcessStatus> STATUSES = new HashSet<>(Arrays.asList(ProcessStatus.values()));
 
     @Override
     public WaitType getType() {
@@ -34,7 +41,12 @@ public class NoneWaitHandler implements ProcessWaitHandler<NoneCondition> {
     }
 
     @Override
-    public NoneCondition process(UUID instanceId, NoneCondition waits) {
+    public Set<ProcessStatus> getProcessStatuses() {
+        return STATUSES;
+    }
+
+    @Override
+    public NoneCondition process(UUID instanceId, ProcessStatus processStatus, NoneCondition waits) {
         return null;
     }
 }